+++ packages/apollo-cache-inmemory/src/queryKeyMaker.ts
@@ -1,7 +1,9 @@
 import { CacheKeyNode } from "./optimism";
 import { DocumentNode, SelectionSetNode, FragmentSpreadNode, FragmentDefinitionNode } from "graphql";
+import { QueryDocumentKeys } from "graphql/language/visitor";
 
 const CIRCULAR = Object.create(null);
+const objToStr = Object.prototype.toString;
 
 export class QueryKeyMaker {
   private perQueryKeyMakers = new Map<DocumentNode, PerQueryKeyMaker>();
@@ -98,21 +100,58 @@ class PerQueryKeyMaker {
   private lookupArray(array: any[]): object {
     const elements = array.map(this.lookupAny, this);
     return this.cacheKeyRoot.lookup(
-      Array.prototype,
+      objToStr.call(array),
       this.cacheKeyRoot.lookupArray(elements),
     );
   }
 
   private lookupObject(object: { [key: string]: any }): object {
-    const keys = Object.keys(object);
-    const locIndex = keys.indexOf("loc");
-    if (locIndex >= 0) keys.splice(locIndex, 1);
-    keys.sort();
+    const keys = safeSortedKeys(object);
     const values = keys.map(key => this.lookupAny(object[key]));
     return this.cacheKeyRoot.lookup(
-      Object.getPrototypeOf(object),
+      objToStr.call(object),
       this.cacheKeyRoot.lookupArray(keys),
       this.cacheKeyRoot.lookupArray(values),
     );
   }
+}
+
+const queryKeyMap: {
+  [key: string]: { [key: string]: boolean }
+} = Object.create(null);
+
+Object.keys(QueryDocumentKeys).forEach(parentKind => {
+  const childKeys = queryKeyMap[parentKind] = Object.create(null);
+
+  (QueryDocumentKeys as {
+    [key: string]: any[]
+  })[parentKind].forEach(childKey => {
+    childKeys[childKey] = true;
+  });
+
+  if (parentKind === "FragmentSpread") {
+    // A custom key that we include when looking up FragmentSpread nodes.
+    childKeys["fragment"] = true;
+  }
+});
+
+function safeSortedKeys(object: { [key: string]: any }): string[] {
+  const keys = Object.keys(object);
+  const keyCount = keys.length;
+  const knownKeys = typeof object.kind === "string" && queryKeyMap[object.kind];
+
+  // Remove unknown object-valued keys from the array, but leave keys with
+  // non-object values untouched.
+  let target = 0;
+  for (let source = target; source < keyCount; ++source) {
+    const key = keys[source];
+    const value = object[key];
+    const isObjectOrArray = value !== null && typeof value === "object";
+    if (! isObjectOrArray || ! knownKeys || knownKeys[key] === true) {
+      keys[target++] = key;
+    }
+  }
+  keys.length = target;
+
+  return keys.sort();
 }